# api/trustgraph.openapi.yaml
openapi: 3.1.0
info:
  title: TrustGraph API
  version: 0.2.0
  description: |
    Trust, safety, and verification service for ConnectionSphere.

    Planes:
      - Plane A: first-party signals and trust assessment.
      - Plane B: consented verification (LinkedIn, government ID, liveness,
        image). Every Plane B read/write requires an active consent.
      - Plane C: investigation tools, restricted to authenticated investigators.

    Auth:
      - Plane A assessment endpoints are called service-to-service.
      - Plane B endpoints are subject-scoped and reached by the app on behalf
        of the subject; consent is enforced server-side per operation.
      - `/v1/admin/*` and `/v1/metrics/*` require `Authorization: Bearer
        $ADMIN_TOKEN`.
      - `/v1/investigations/*` require `Authorization: Bearer
        $INVESTIGATOR_TOKEN` (or `$ADMIN_TOKEN` for supervisor-level access),
        plus an optional `X-Investigator-Actor` header naming the human.

    When a vendor integration is not configured, its endpoint returns
    `503 not_configured`. It never returns a passing result it did not obtain.
servers:
  - url: http://localhost:8081
    description: Local development (via ConnectionSphere Docker network)
  - url: https://trustgraph-api.onrender.com
    description: Render production
...
paths:
...
  /v1/verification/subjects/{csUserId}/consents:
    get:
      operationId: listConsents
      summary: List a subject's consent records
      tags: [Verification (Plane B)]
      parameters:
        - $ref: "#/components/parameters/CsUserId"
      responses:
        "200":
          description: Consent records
          content:
            application/json:
              schema:
                type: object
                properties:
                  consents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Consent"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: grantConsent
      summary: Grant consent for a verification type
      tags: [Verification (Plane B)]
      parameters:
        - $ref: "#/components/parameters/CsUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentGrantRequest"
      responses:
        "201":
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consent"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/verification/subjects/{csUserId}/consents/{consentType}:
    delete:
      operationId: withdrawConsent
      summary: Withdraw consent and delete the associated Plane B data
      tags: [Verification (Plane B)]
      description: |
        Withdrawing consent erases the data collected under that purpose and
        marks the consent row withdrawn. The erase is scoped to the purpose:
        withdrawing `image_verification` does not remove a subject's LinkedIn
        profile, and a subject's other consents stay granted. Use
        `DELETE /v1/verification/subjects/{csUserId}` to erase everything.

        The consent row itself is retained (status `withdrawn`) as an audit
        trail of what was once granted; it is not returned as active consent.
      parameters:
        - $ref: "#/components/parameters/CsUserId"
        - name: consentType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ConsentType"
      responses:
        "200":
          description: Consent withdrawn and the data held under it deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/verification/subjects/{csUserId}/badges:
    get:
      operationId: getBadges
      summary: Get a subject's earned verification badges
      tags: [Verification (Plane B)]
      parameters:
        - $ref: "#/components/parameters/CsUserId"
      responses:
        "200":
          description: Badges
          content:
            application/json:
              schema:
                type: object
                properties:
                  subjectId:
                    type: string
                    format: uuid
                  badges:
                    type: array
                    items:
                      $ref: "#/components/schemas/Badge"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/verification/subjects/{csUserId}:
    delete:
      operationId: deleteVerification
      summary: Delete all Plane B verification data for a subject
      description: |
        The full-erasure path for Plane B: removes every verification record,
        LinkedIn profile and token, employment history, and consent-linked data
        for the subject. Use the purpose-scoped consent withdrawal when only one
        purpose should be erased.
      tags: [Verification (Plane B)]
      parameters:
        - $ref: "#/components/parameters/CsUserId"
      responses:
        "204":
          description: Data deleted; no response body
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    CsUserId:
      name: csUserId
      in: path
      required: true
      description: ConnectionSphere user ID
      schema:
        type: string
        example: usr_12345
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested subject, case, or consent does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    ServiceAccountHeader:
      type: apiKey
      in: header
      name: X-Service-Account
      description: ConnectionSphere service account identifier
    AdminBearer:
      type: http
      scheme: bearer
      description: "Set via ADMIN_TOKEN. Grants supervisor-level Plane C access."
    InvestigatorBearer:
      type: http
      scheme: bearer
      description: "Set via INVESTIGATOR_TOKEN. Optional X-Investigator-Actor header names the human."
  schemas:
...
    ConsentType:
      type: string
      enum: [linkedin_oauth, government_id, liveness, image_verification]
    ConsentGrantRequest:
      type: object
      required: [consentType]
      properties:
        consentType:
          $ref: "#/components/schemas/ConsentType"
        policyVersion:
          type: string
          description: Version of the consent policy the subject accepted
        termsAccepted:
          type: object
          additionalProperties: true
          description: Opaque record of the specific terms the subject accepted
    Consent:
      type: object
      description: |
        A consent record. Withdrawal is a soft state change rather than a
        delete, so withdrawn records remain queryable for audit.
      properties:
        consentId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        plane:
          type: string
          enum: [B]
        consentType:
          $ref: "#/components/schemas/ConsentType"
        consentStatus:
          type: string
          enum: [granted, withdrawn, expired]
        policyVersion:
          type: string
        termsAccepted:
          type: object
          additionalProperties: true
        grantedAt:
          type: string
          format: date-time
        withdrawnAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Badge:
      type: object
      description: A user-visible trust badge derived from verification state.
      properties:
        type:
          type: string
          example: government_id
        state:
          type: string
          enum: [verified, self_reported, none]
        label:
          type: string
          example: Government ID verified
        detail:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string